Added an ability to unregister single subscriptions.
It is needed when the subscriber is an object with the lifecycle and it
dies in some point of time. The most notable example of such subscribers
are React components. Flux popularized the flow where subscriptions to
event buses are within the boundary of the component itself - thus it is
often needed.