Let's keep our build releases inside GitHub's releases
These files should be easy to build (and they are), so there is no point of
keeping those files inside repository tree.

This is just a common practice I saw in another projects. We do GitHub and NPM
releases from time to time, so there is no need to keep the redundancy directly
in repository.